refactor(DEV-11277): rename DropIn files
